+++ test/Converter.Core.UnitTests/GlobalUsings.cs
@@ -1,0 +1,5 @@
+﻿// Global using directives
+
+global using FluentAssertions;
+global using NSubstitute;
+global using Xunit;